Refactor API Structure and Implement WebSocket for Processing Updates
- Renamed `deps.py` to `dependencies.py` for clarity and consistency across the API structure.
- Created a new WebSocket endpoint at `/api/ws/processing` to stream real-time processing updates, allowing the frontend to receive live status updates during item processing.
- Introduced a `ProcessingConnectionManager` to manage WebSocket connections and handle message broadcasting with backpressure policies.
- Updated the `ProcessingUpdate` model to standardize the structure of processing event messages.
- Enhanced the `main.py` file to register the new WebSocket router and manage the lifecycle of the processing queue and WebSocket connections.
- Added comprehensive tests for the WebSocket functionality to ensure robust handling of connections and message delivery.

+++ python-backend/src/api/routes/health.py
@@ -7,10 +7,10 @@ import aiosqlite
 from fastapi import APIRouter, Depends, status
 from fastapi.responses import JSONResponse
 
-from ..config import get_app_version
-from ..db.models import ComponentCheck, HealthResponse
-from ..providers import OllamaHealthResponse, OllamaProvider
-from .deps import get_db_connection, get_ollama_provider
+from ...config import get_app_version
+from ...db.models import ComponentCheck, HealthResponse
+from ...providers import OllamaHealthResponse, OllamaProvider
+from ..dependencies import get_db_connection, get_ollama_provider
 
 router = APIRouter(tags=["health"])
 
@@ -110,7 +110,7 @@ async def check_ollama_health(
     start = time.perf_counter()
     try:
         models = await provider.list_models()
-        latency_ms = (time.perf_counter() - start) * 1000
+        latency_ms = int((time.perf_counter() - start) * 1000)
         return OllamaHealthResponse(
             status="healthy",
             base_url=provider.base_url,
